requirements.txt
Flask==2.3.3
openai
newsapi-python
datetime
schedule
gunicorn